+++ lib/hexapdf/object.rb
@@ -284,12 +284,14 @@ module HexaPDF
       obj
     end
 
-    # Caches and returns the given value or, if not provided, the one returned by the block under
-    # the given cache key. If there is already a cached value for the key, it is just returned.
+    # Caches and returns the given +value+ or the value of the block under the given cache key. If
+    # there is already a cached value for the key and +update+ is +false+, it is just returned.
+    #
+    # Set +update+ to +true+ to force an update of the cached value.
     #
     # This uses Document#cache internally.
-    def cache(key, value = nil, &block)
-      document.cache(@data, key, value, &block)
+    def cache(key, value = Document::UNSET, update: false, &block)
+      document.cache(@data, key, value, update: update, &block)
     end
 
     # Returns +true+ if there is a cached value for the given key.